Moved OrderList to Grocery inner class.
